Show custom dialog based on stung or paused
We now show a slightly different dialog depending on if the player was
stung or hit the pause button. If they hit the pause button, we have mute
option in the dialog.  It also has a different title so it doesn't say the
player was stung.  We will want to make this look pretty.